[#25] Add `run` subcommand for one-off env usage
I also made the variable/application command arg required. Previously it was optional, but not passing it would spit out an error with a suggestion. The suggestion was nice, but I think it made the usage string confusing. In the future we can change it back to optional and offer an additional TUI to select var/application (before the value/profile selection).